Extract wikilinks.py from legacy.py
